Replace separate classes with factory methods
