+++ README.md
@@ -21,38 +21,37 @@ Let's assume that your configuration is formatted as JSON:
 }
 ```
 
-To load and automatically refresh the configuration from S3, create a case class
-that represents your configuration (e.g. `FrozzlerConfiguration` in the example
-below). Then, call `DynamicConfigurationFromS3`, passing in the bucket and key
-at which your configuration file is located and a method for converting from the
-string content of your configuration to a `Try[FrozzlerConfiguration]`.
+To load and automatically refresh the configuration from S3, create a case
+class that represents your configuration (e.g. `FrozzlerConfiguration` in the
+example below). Then, call `DynamicConfiguration.fromS3`, passing in the bucket
+and key at which your configuration file is located and a method for converting
+from the string content of your configuration to a
+`Try[FrozzlerConfiguration]`.
 
-`DynamicConfigurationFromS3` will return a `DynamicConfiguration` object with a
-`currentConfiguration` method. This returns an option with either the current
+`DynamicConfiguration.fromS3` will return a `DynamicConfiguration` object with
+a `currentConfiguration` method. This returns an option with either the current
 configuration, or `None` if the configuration is not loaded yet.
 
 ```scala
-import com.asidatascience.configuration.{DynamicConfigurationFromS3, RefreshOptions}
-
-import scala.util.Try
-import org.json4s._
-
 import scala.concurrent.duration._
+import scala.util.Try
 
 import com.amazonaws.regions.Regions
 import com.amazonaws.services.s3.AmazonS3ClientBuilder
+import com.asidatascience.configuration.{DynamicConfiguration, RefreshOptions}
+import org.json4s._
 
 final case class FrozzlerConfiguration(model: String)
 
 class WidgetFrozzler(
-  configurationS3Bucket: String,
-  configurationS3Key: String
+    configurationS3Bucket: String,
+    configurationS3Key: String
 ) {
 
   implicit val actorSystem = ActorSystem()
 
   private def parseConfiguration(content: String) = {
-    // parse the contents of the configuration file
+    // Parse the contents of the configuration file.
     val contentAsJson = JsonMethods.parse(content)
     val JString(model) = (contentAsJson \ "widget-model")
     FrozzlerConfiguration(model)
@@ -64,16 +63,21 @@ class WidgetFrozzler(
   )
 
   val s3Client = AmazonS3ClientBuilder
-    .standard().withRegion(Regions.EU_WEST_1).build
+    .standard()
+    .withRegion(Regions.EU_WEST_1)
+    .build
 
-  lazy val configurationService = DynamicConfigurationFromS3[FrozzlerConfiguration](
-    s3Client,
-    configurationS3Bucket,
-    configurationS3Key,
-    refreshOptions
-  ){ contents => Try { parseConfiguration(contents) } }
+  lazy val configurationService =
+    DynamicConfiguration.fromS3[FrozzlerConfiguration](
+      s3Client,
+      configurationS3Bucket,
+      configurationS3Key,
+      refreshOptions
+    ) { contents =>
+      Try { parseConfiguration(contents) }
+    }
 
-  def frozzleWidgets = {
+  def frozzleWidgets =
     configurationService.currentConfiguration match {
       case Some(configuration) =>
         val currentModel = configuration.model
@@ -81,7 +85,6 @@ class WidgetFrozzler(
       case None =>
         println("Configuration not ready")
     }
-  }
 }
 ```
