+++ src/main/scala/com/asidatascience/configuration/RefreshOptions.scala
@@ -2,7 +2,7 @@ package com.asidatascience.configuration
 
 import scala.concurrent.duration._
 
-case class RefreshOptions(
+final case class RefreshOptions(
     initialDelay: FiniteDuration = 0.millis,
     updateInterval: FiniteDuration = 5.minutes
 )
